Delete MultiWorker; flake8 & pylint cleanups.
MultiWorker tried to use :mod:`multiprocessing` as a framework to run
one process per task, but its odd use of threads and forking results in
trouble if there is any locking inside the parent process (as, for instance,
happens inside :mod:`logging`).  ForkWorker is more soundly based on
Unix primitives (os.fork()) and is pretty proven.