Fix unrestrained work unit creation for continuous work specs.
The new `background` work spec operates by:

  1. Configuring a work spec that is continuous and runs
     at an interval of 120 seconds.
  2. Does some bounded scanning at each work unit execution,
     and saves the state at which the scanning stopped in the
     work unit state by creating a new work unit upon
     completion.

As I understand it, this work unit should *always* run at intervals
of 120 seconds. That is, under no circumstances should it wind up in
a loop where work units for this work spec are executed immediately,
regardless of whether the work unit actually creates new work units
for itself.

In the days of yore, it seems continuous work units on an interval
were executed by coordinated via something called a "synthetic"
work unit. Therefore, coordinated takes care of work unit creation
and execution and the world is sane. But in the background work spec,
coordinated is no longer able to create work units since work units
depend on some custom state for the work spec. This is a *useful*
technique, so it should be supported.

Therefore, this commit modifies the `get_work` method on the `WorkSpec`
class to only produce work for continuous/interval work units only if
the appropriate amount of time has expired, even if there is a work
unit available in the queue.