Try to fix del_work_units on SQLite.
The work spec implementation was trying to pass along an extra state
parameter to the underlying storage, only used for available jobs.

To trigger this you needed to have SQLite set up underneath coordinated,
and to call del_work_units() with specific work unit keys to delete
(not "delete all available" or "delete everything") and not limit the
state to something not-available.