Allow coordinate_worker to run as non-root (again).
python-daemon 2.1.0 changes to call os.initgroups() to set the complete
group list, which (unlike os.setgid()) requires root privileges.  There is
an option to go back to the older setgid() behavior; so require a current
version of python-daemon but tell it to act like the old version.  (Diffeo
internal issue DIFFEO-2375.)